YYC-109: split agent.rs into agent/ submodules
agent.rs grew to ~1800 lines covering construction, iteration loop,
tool dispatch, summary helpers, session lifecycle, and provider plumbing.
Split into focused submodules so each surface is independently testable.

* agent/mod.rs (394 lines) — Agent struct, ModelSelection, constructors,
  state accessors, save_messages, for_test, flatten_for_message helper.
* agent/dispatch.rs — dispatch_tool + summarize_tool_args /
  summarize_tool_result / preview_output / elided_lines.
* agent/session.rs — start_session, end_session, resume_session,
  continue_last_session, fork_session.
* agent/provider.rs — switch_model, switch_provider,
  restore_persisted_provider, fetch_catalog_for, resolve_model_selection,
  available_models.
* agent/run.rs — run_prompt + run_prompt_stream.
* agent/tests.rs — unit tests (agent integration + helper coverage).

Fields on Agent move to pub(in crate::agent) so sibling submodules can
reach them while staying invisible outside the agent module.

Public API unchanged. cargo test --all-features --lib passes (242 tests).
